First version of HomeScreen implemented with initial drawer navigation
Added TransportationTypeRow component with all the different buttons for each type of transportation.
Also added the Favourite list with a dummy data of few buses in leuven.
Added the first inital drawer screen with the drawer navigation implemented to the homeScreen.
And to finalise tweak the styles so the components fit nice into the HomeScreen.